Better separation between schemas and serializers
